moved AutoMapper from ADT Repo to SDK
