+++ src/AAS ADT SDK Tests/AAS ADT SDK Tests.csproj
@@ -10,6 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+	<PackageReference Include="AutoMapper" Version="12.0.0" />
 	<PackageReference Include="FluentAssertions" Version="6.9.0" />
 	<PackageReference Include="Microsoft.Extensions.Azure" Version="1.5.0" />
 	<PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
