+++ src/AAS ADT SDK Tests/AutoMapper/AdtIec61360ProfileTests.cs
@@ -1,10 +1,10 @@
-﻿using AAS.ADT.Models;
-using AAS.API.Repository.Adt.AutoMapper;
+﻿using AAS.ADT.AutoMapper;
+using AAS.ADT.Models;
 using AasCore.Aas3_0_RC02;
 using AutoMapper;
 using FluentAssertions;
 
-namespace AAS.API.Repository.Adt.Tests
+namespace AAS.ADT.Tests.AutoMapper
 {
     [TestClass]
     public class AdtIec61360ProfileTests
