+++ src/AAS ADT SDK/AAS ADT SDK.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="AasCore.Aas3_0_RC02" Version="1.1.0" />
+    <PackageReference Include="AutoMapper" Version="12.0.0" />
     <PackageReference Include="Azure.DigitalTwins.Core" Version="1.4.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.2" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
