+++ src/AAS ADT SDK/AutoMapper/AdtIec61360Profile.cs
@@ -1,7 +1,8 @@
-﻿using AAS.ADT.Models;
+﻿using System.Collections.Generic;
+using AAS.ADT.Models;
 using AasCore.Aas3_0_RC02;
 
-namespace AAS.API.Repository.Adt.AutoMapper
+namespace AAS.ADT.AutoMapper
 {
     public class AdtIec61360Profile : AdtBaseProfile
     {
